+++ cli/docs/changelog.md
@@ -22,6 +22,8 @@
 ### Fixes
 * Fixed a bug where pods would return before being started
 * Upgraded to container-engine-lib 0.25.0 (and downstream deps), which fix several issues on Kubernetes
+* Bumped the number of retry polls in the module tests up to 20 to try and remediate a timeout while waiting for the datastore service to come up that we were seeing
+* Fixed the config getting serialized strangely
 * Fixed a bug where Kubrnetes service updates would sometimes fail
 
 ### Breaking Changes